Refactor and remove some tests

+++ tests/test_features.py
@@ -19,9 +19,10 @@ def test_fit_vectorizer_returns_vectorizer_and_matrix() -> None:
     
     reviews = ["great hotel", "bad room", "great stay"]
     vec, matrix = fit_vectorizer(reviews)
+
     assert isinstance(vec, TfidfVectorizer)
     assert matrix.shape == (3, matrix.shape[1])
-    assert matrix.dtype in (np.float32, np.float64)
+    assert matrix.dtype in (float, float)
     assert not np.any(np.isnan(matrix))
 
 
@@ -31,24 +32,28 @@ def test_transform_uses_fitted_vectorizer() -> None:
     train = ["good", "bad", "nice"]
     vec, train_matrix = fit_vectorizer(train)
     test_matrix = transform(vec, ["good nice"])
+
     assert test_matrix.shape[0] == 1
     assert test_matrix.shape[1] == train_matrix.shape[1]
 
 
 def test_save_and_load_vectorizer(tmp_path_in_project: Path) -> None:
     """Saving and loading round-trips the vectorizer."""
+    
     paths = Paths(project_root=tmp_path_in_project)
     paths.artifacts.mkdir(parents=True, exist_ok=True)
 
     reviews = ["one two", "two three", "three four"]
     vec, _ = fit_vectorizer(reviews)
     save_vectorizer(vec, paths)
+
     assert (tmp_path_in_project / "artifacts" / "tfidf.joblib").exists()
+
     loaded = load_vectorizer(paths)
+
     assert loaded.vocabulary_ == vec.vocabulary_
+
     matrix_orig = transform(vec, ["one three"])
     matrix_loaded = transform(loaded, ["one three"])
+
     np.testing.assert_array_almost_equal(matrix_orig, matrix_loaded)
-
-
-
